Add TestClient and simplify API by /publish
Before a /connect and /disconnect was needed and corrupted nodes could not be detected.
Within the new /connect a child is automatically added and it is removed from the list if it does not poll for more than 2 minutes
The TestClient is an example for other software to connect to AAS Connect or to port the client to other programming languages.